+++ src/main/scala/gitbucket/core/model/Profile.scala
@@ -45,7 +45,7 @@ trait CoreProfile
     with Profile
     with AccessTokenComponent
     with AccountComponent
-    with ActivityComponent
+    with ActivityComponent // ActivityComponent has been deprecated, but keep it for binary compatibility
     with CollaboratorComponent
     with CommitCommentComponent
     with CommitStatusComponent