+++ src/test/scala/gitbucket/core/service/MergeServiceSpec.scala
@@ -6,15 +6,15 @@ import gitbucket.core.util.GitSpecUtil._
 import org.eclipse.jgit.api.Git
 import org.eclipse.jgit.lib._
 import org.eclipse.jgit.revwalk._
-import org.scalatest.FunSpec
+import org.scalatest.funspec.AnyFunSpec
 
 import java.io.File
 import scala.util.Using
 
-class MergeServiceSpec extends FunSpec {
+class MergeServiceSpec extends AnyFunSpec {
   val service = new MergeService with AccountService with ActivityService with IssuesService with LabelsService
   with MilestonesService with RepositoryService with PrioritiesService with PullRequestService with CommitsService
-  with WebHookPullRequestService with WebHookPullRequestReviewCommentService {}
+  with WebHookPullRequestService with WebHookPullRequestReviewCommentService with RequestCache {}
   val branch = "master"
   val issueId = 10
   def initRepository(owner: String, name: String): File = {